requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
bleach==6.3.0
python-slugify==8.0.1
email-validator==2.1.1
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9